requirements.txt
py-cord~=2.4.1
python-dotenv~=1.0.0
SQLAlchemy~=2.0.21
requests~=2.31.0